--- UnstoppableWallet/UnstoppableWallet/Modules/Send/FeeSettings/SendFeeSettingsService.swift
@@ -1,22 +1,0 @@
-import Foundation
-import RxSwift
-import RxRelay
-
-class SendFeeSettingsService {
-    private let disposeBag = DisposeBag()
-
-    private let defaultFeeRateRelay = BehaviorRelay<Bool>(value: true)
-
-}
-
-extension SendFeeSettingsService {
-
-    var defaultFeeRateObservable: Observable<Bool> {
-        defaultFeeRateRelay.asObservable()
-    }
-
-    func reset() {
-        // Set recommended feeRate here
-    }
-
-}
