--- UnstoppableWallet/UnstoppableWallet/Modules/Send/FeeSettings/SendFeeSettingsViewController.swift
@@ -1,200 +1,0 @@
-import UIKit
-import ThemeKit
-import SnapKit
-import SectionsTableView
-import RxSwift
-import RxCocoa
-import ComponentKit
-
-class SendFeeSettingsViewController: ThemeViewController {
-    let disposeBag = DisposeBag()
-
-    private let viewModel: SendFeeSettingsViewModel
-
-    private let tableView = SectionsTableView(style: .grouped)
-    let bottomWrapper = BottomGradientHolder()
-
-    private let feeViewModel: SendFeeViewModel
-    private let feeCautionViewModel: SendFeeWarningViewModel
-    private let amountCautionViewModel: SendFeeSettingsAmountCautionViewModel
-
-    private let feeCell: FeeCell
-    private let feeRateCell = BaseThemeCell()
-    private let feeWarningCell = TitledHighlightedDescriptionCell()
-    private let feeAmountErrorCell = TitledHighlightedDescriptionCell()
-
-    private let doneButton = PrimaryButton()
-
-    private var loaded = false
-
-    init(viewModel: SendFeeSettingsViewModel, feeViewModel: SendFeeViewModel, feeCautionViewModel: SendFeeWarningViewModel, amountCautionViewModel: SendFeeSettingsAmountCautionViewModel) {
-        self.viewModel = viewModel
-        self.feeViewModel = feeViewModel
-        self.feeCautionViewModel = feeCautionViewModel
-        self.amountCautionViewModel = amountCautionViewModel
-
-        feeCell = FeeCell(viewModel: feeViewModel, title: "fee_settings.fee".localized)
-
-        super.init()
-    }
-
-    required init?(coder aDecoder: NSCoder) {
-        fatalError("init(coder:) has not been implemented")
-    }
-
-    override func viewDidLoad() {
-        super.viewDidLoad()
-
-        title = "fee_settings".localized
-        navigationItem.rightBarButtonItem = UIBarButtonItem(title: "button.done".localized, style: .done, target: self, action: #selector(onTapDone))
-        navigationItem.leftBarButtonItem = UIBarButtonItem(title: "button.reset".localized, style: .plain, target: self, action: #selector(onTapReset))
-
-        view.addSubview(tableView)
-        tableView.snp.makeConstraints { maker in
-            maker.leading.top.trailing.equalToSuperview()
-        }
-
-        tableView.backgroundColor = .clear
-        tableView.separatorStyle = .none
-        tableView.delaysContentTouches = false
-
-        tableView.registerCell(forClass: TitledHighlightedDescriptionCell.self)
-        tableView.sectionDataSource = self
-
-        feeRateCell.set(backgroundStyle: .lawrence, isFirst: true, isLast: false)
-        sync(feeSliderViewItem: nil)
-
-        view.addSubview(bottomWrapper)
-        bottomWrapper.snp.makeConstraints { maker in
-            maker.top.equalTo(tableView.snp.bottom).offset(-CGFloat.margin16)
-            maker.leading.trailing.equalToSuperview()
-            maker.bottom.equalTo(view.safeAreaLayoutGuide)
-        }
-
-        bottomWrapper.addSubview(doneButton)
-        doneButton.snp.makeConstraints { maker in
-            maker.top.equalToSuperview().inset(CGFloat.margin32)
-            maker.leading.trailing.equalToSuperview().inset(CGFloat.margin24)
-            maker.bottom.equalToSuperview().inset(CGFloat.margin16)
-        }
-
-        doneButton.set(style: .yellow)
-        doneButton.setTitle("button.done".localized, for: .normal)
-        doneButton.addTarget(self, action: #selector(onTapDone), for: .touchUpInside)
-
-        subscribe(disposeBag, feeCautionViewModel.cautionDriver) { [weak self] in
-            self?.handle(cell: self?.feeWarningCell, caution: $0)
-        }
-        subscribe(disposeBag, amountCautionViewModel.amountCautionDriver) { [weak self] in
-            self?.handle(cell: self?.feeAmountErrorCell, caution: $0)
-        }
-
-        subscribe(disposeBag, viewModel.resetButtonActiveDriver) { [weak self] active in
-            self?.navigationItem.leftBarButtonItem?.isEnabled = active
-        }
-
-        tableView.buildSections()
-
-        loaded = true
-    }
-
-    @objc private func onTapDone() {
-        dismiss(animated: true)
-    }
-
-    @objc private func onTapReset() {
-        viewModel.reset()
-    }
-
-    private func handle(cell: TitledHighlightedDescriptionCell?, caution: TitledCaution?) {
-        guard let cell = cell else {
-            return
-        }
-        cell.isVisible = caution != nil
-
-        if let caution = caution {
-            cell.bind(caution: caution)
-        }
-
-        if loaded {
-            UIView.animate(withDuration: 0.15) {
-                self.tableView.beginUpdates()
-                self.tableView.endUpdates()
-            }
-        }
-    }
-
-    private func sync(feeSliderViewItem: FeeViewItem?) {
-        CellBuilderNew.buildStatic(cell: feeRateCell, rootElement: .hStack(
-                tableView.universalImage24Elements(title: .subhead2("fee_settings.fee_rate".localized), value: .subhead1(feeSliderViewItem?.description)))
-        )
-    }
-
-    private func openInfo(title: String, description: String) {
-        guard feeViewModel.showInfoIcon else {
-            return
-        }
-
-        let viewController = BottomSheetModule.description(title: title, text: description)
-        present(viewController, animated: true)
-    }
-
-    private func reloadTable() {
-        if loaded {
-            UIView.animate(withDuration: 0.15) {
-                self.tableView.beginUpdates()
-                self.tableView.endUpdates()
-            }
-        }
-    }
-
-}
-
-extension SendFeeSettingsViewController: SectionsDataSource {
-
-    func buildSections() -> [SectionProtocol] {
-        let feeSections: [SectionProtocol] = [
-            Section(
-                    id: "fee",
-                    headerState: .margin(height: .margin12),
-                    rows: [
-                        StaticRow(
-                                cell: feeCell,
-                                id: "fee",
-                                height: .heightCell48,
-                                autoDeselect: true,
-                                action: { [weak self] in
-                                    self?.openInfo(title: "fee_settings.fee".localized, description: "fee_settings.fee.info".localized)
-                                }
-                        )
-                    ]
-            )
-        ]
-
-        let cautionsSections: [SectionProtocol] = [
-            Section(
-                    id: "cautions",
-                    headerState: .margin(height: .margin12),
-                    rows: [
-                        StaticRow(
-                                cell: feeWarningCell,
-                                id: "fee-caution",
-                                dynamicHeight: { [weak self] containerWidth in
-                                    self?.feeWarningCell.cellHeight(containerWidth: containerWidth) ?? 0
-                                }
-                        ),
-                        StaticRow(
-                                cell: feeAmountErrorCell,
-                                id: "amount-error",
-                                dynamicHeight: { [weak self] containerWidth in
-                                    self?.feeAmountErrorCell.cellHeight(containerWidth: containerWidth) ?? 0
-                                }
-                        )
-                    ]
-            )
-        ]
-
-        return feeSections + cautionsSections
-    }
-
-}
